ftests: Add --container and --no-container flags
In preparation of running the tests outside of a
container, add two flags to ftests.py: --container
and --no-container.  As the names imply, these flags
are used to specify whether the tests should be
containerized or not.

Note1: Some tests (cgrulesengd) will only work
       in a non-containerized environment.

Note2: Running the tests in a non-containerized
       environment could be destructive to a
       system as cgroup changes could be made that
       adversely affect the entire system.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>