ftests: Update test 016 to match controller logic
The cgroup_controller structure was updated to store the
controller's version.  This change slightly affected the
behavior of cgget when an invalid controller is passed in
via the '-r' flag.  Previously an erroneous controller
was not observed by cgget, until cgget tried to read the
setting.  Now with the controller version being stored in
the struct, libcgroup knows that the setting is invalid
much earlier.

Update test 016 to be compatible with the new behavior
outlined above.  Now when a user requests an invalid
setting via '-r', libcgroup will raise an error code
of 92.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>